SqlConnectivity Test Multi Driver
We usually have service requests that our customers using different drivers need to know the connection latency, execution time spent or validate the connection.

For this reason, I developed this small PowerShell script  (based on initiative of AzureSQLConnectivityChecker in multiple languages) which will basically connect to the database using 4 drivers/providers and test this.